added schema Draft7 form validation; narrowed bug to validation error
