+++ nwb_conversion_tools/schemas/metadata_schema.json
@@ -11,7 +11,7 @@
     "NWBFile": {
       "type": "object",
       "additionalProperties": false,
-      "required": ["session_start_time", "keywords"],
+      "required": [],
       "properties": {
         "keywords": {
           "description": "Terms to search over",
